Render authenticated data on first page load (#136)
This commit adds cookie forwarding to requests in getInitialProps to remove the need to send AJAX requests once the page loads on the client. User data and individual club data will be loaded in on the NextJS server, and sent in the initial HTML payload to the client.

This gets rid of any flicker on page load.